✨ Adds date field to posts for news items
Adds a date field to the posts table to allow for displaying the date of news items and ordering them by date.

This change introduces a new 'date' field to the posts table, which is used to store the publication date of news items. This date is then used to display the date alongside the news item and to determine the order in which news items are displayed.

Also includes translations for the new field in English and Dutch.